requirements.txt
openai
pandas
beautifulsoup4
requests
tiktoken
python-dotenv